+++ internal/go_repository_tools_srcs.bzl
@@ -181,6 +181,11 @@ GO_REPOSITORY_TOOLS_SRCS = [
     Label("//v2/testtools:BUILD.bazel"),
     Label("//v2/testtools:config.go"),
     Label("//v2/testtools:files.go"),
+    Label("//v2/walk:BUILD.bazel"),
+    Label("//v2/walk:cache.go"),
+    Label("//v2/walk:config.go"),
+    Label("//v2/walk:dirinfo.go"),
+    Label("//v2/walk:walk.go"),
     Label("//walk:BUILD.bazel"),
     Label("//walk:cache.go"),
     Label("//walk:config.go"),
